Safari-611.1.10/LayoutTests/fast/visual-viewport/ios/min-scale-greater-than-one-expected.txt

PASS Math.round(boundingClientRect.top) is expectedTop
Scrolling to bottom
PASS window.scrollY is expectedTop
PASS Math.round(boundingClientRect.top) is expectedTop
Tap dispatched
Target clicked
PASS successfullyParsed is true

TEST COMPLETE
